pyproject.toml
[project]
name = "api-consulta-sigp"
version = "0.1.0"
description = "Sistema de automação para inserção automatica de documentos"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "async-generator==1.10",
    "attrs==25.4.0",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "fastapi==0.128.0",
    "h11==0.16.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "keyboard==0.13.5",
    "mypy-extensions==1.1.0",
    "outcome==1.3.0.post0",
    "packaging==26.0",
    "pexpect==4.9.0",
    "pillow==12.1.0",
    "psutil==7.2.2",
    "ptyprocess==0.7.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pypdf==6.7.0",
    "pyperclip==1.11.0",
    "pysocks==1.7.1",
    "python-dotenv==1.2.1",
    "reportlab==4.4.9",
    "requests==2.32.5",
    "selenium==4.40.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "starlette==0.50.0",
    "trio==0.32.0",
    "trio-typing==0.10.0",
    "trio-websocket==0.12.2",
    "types-certifi==2021.10.8.3",
    "types-urllib3==1.26.25.14",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.40.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.9.0",
    "wsproto==1.3.2",
    "zipp==3.23.0",
]
